fix: move interface to general.ts
