+++ src/routes/graphql/index.ts
@@ -1,23 +1,7 @@
-/* eslint-disable @typescript-eslint/no-unsafe-argument */
-/* eslint-disable @typescript-eslint/no-unsafe-return */
 import { FastifyPluginAsyncTypebox } from '@fastify/type-provider-typebox';
 import { createGqlResponseSchema, gqlResponseSchema, result } from './schemas.js';
 import { graphql, parse, validate } from 'graphql';
 import depthLimit from 'graphql-depth-limit';
-
-export type User = {
-  id: string;
-  name: string;
-  balance: number;
-  userSubscribedTo?: {
-    subscriberId: string;
-    authorId: string;
-  }[];
-  subscribedToUser?: {
-    subscriberId: string;
-    authorId: string;
-  }[];
-};
 
 const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
   fastify.route({
@@ -34,8 +18,6 @@ const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
 
       const source = req.body.query;
       const { variables } = req.body;
-
-      console.log('var:', variables);
 
       const validationErrors = validate(result, parse(source), [depthLimit(5)]);
